BFF-637 Separated GiftAid from registration process
